TP3/requirements.txt
flask==3.1.2
google-cloud-storage==3.5.0
